pipeline/lambda_function/requirements.txt
boto3
openpyxl
pandas
redshift_connector
pytest
pylint
python-dotenv
pyarrow
altair
vega_datasets
rich